+++ CHANGELOG.md
@@ -5,6 +5,7 @@
 * Support setting local variables of primitive type (issue #46)
 * Support Debugger.continueToLocation (issue #47)
 * Support Debugger.pause (issue #49)
+* Pause in certain situations even if breakpoints are disabled (issue #50)
 
 # 0.5.0 (2017-05-28)